+++ qa/lua/run.lua
@@ -91,6 +91,7 @@ local suites = {
     "test_et_boss_ideas",
     "test_et_custom_breed_registrar",
     "test_et_custom_breed_parity",
+    "test_et_issue1497_local_host_fence",
     "test_et_pacing_tick_guard",
     "test_rt_harness_context_absent",
     "test_event_high_intensity_diagnostic",